BE -configuracion de repositorios y servicios de restriction exception and history

+++ backend/src/main/java/com/jt/backend/repositories/ConsultasRepositoryInterface.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jt.backend.models.Consulta;
 
-public interface ConsultasRepositoryInterface extends JpaRepository<Consulta, Integer>{
+public interface ConsultasRepositoryInterface extends JpaRepository<Consulta, Long>{
 
 }
